requirements.txt
astroid==1.4.9
autopep8==1.3.1
coloredlogs==6.0
docopt==0.6.2
humanfriendly==2.4
isort==4.2.5
lazy-object-proxy==1.2.2
mccabe==0.6.1
nose==1.3.7
pluggy==0.4.0
py==1.4.33
pycodestyle==2.3.1
pylint==1.6.5
PyYAML==3.12
six==1.10.0
tox==2.6.0
virtualenv==15.1.0
wrapt==1.10.10
